Move responsibility to load the JavaQuasarBridge library to the client.
While nothing Quasar related will work before a Quasar host is created,
it could in principle be possible that we have some (native) support
code that does not actually call Quasar in the JavaQuasarBridge. If we
then rely on that static initializer of QHost to load the native
library, we might load it too late or not at all.